+++ Modules/CertificateDsc/Examples/Resources/PfxImport/2-MinimalUsage.ps1
@@ -1,0 +1,32 @@
+<#
+    .EXAMPLE
+    Import a PFX into the My store.
+#>
+Configuration Example
+{
+    param
+    (
+        [Parameter()]
+        [System.String[]]
+        $NodeName = 'localhost',
+
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullorEmpty()]
+        [System.Management.Automation.PSCredential]
+        $Credential
+    )
+
+    Import-DscResource -ModuleName CertificateDsc
+
+    Node $AllNodes.NodeName
+    {
+        PfxImport CompanyCert
+        {
+            Thumbprint = 'c81b94933420221a7ac004a90242d8b1d3e5070d'
+            Path       = '\\Server\Share\Certificates\CompanyCert.pfx'
+            Location   = 'LocalMachine'
+            Store      = 'My'
+            Credential = $Credential
+        }
+    }
+}